Add CLI and API endpoints
Changed the previous sandbox CLI tool into a more mature, separate
executable (simply called stc). This, besides doing all the previously
implemented things, also has a couple more config options, as well as a
help screen, and a version info screen, including some info useful for
debugging builds.

An extern "C" endpoint has also been added to pkg_api.h in the julia
frontend. This will be the entry point for the julia-side wrapper pkg.

Also includes a bunch of clang-tidy noticed bugfixes and
recommendations, which means the project should be at 0 warnings again,
even if building with clang-tidy.

Finally, a build_all powershell script has been added to scripts, that
ties together builds across different compilers and cmake configs.

+++ include/frontend/jl/ast.h
@@ -184,7 +184,7 @@ struct VarDecl : public Decl {
         if (value)
             _node_storage |= (1U << 7);
         else
-            _node_storage &= ~(1U << 7);
+            _node_storage &= static_cast<uint8_t>(~(1U << 7));
     }
 
     SAME_NODE_KIND_DEF(NodeKind::VarDecl)
